+++ models/User.js
@@ -5,7 +5,7 @@ const schema = new Schema ({
     email: { type: String, required: true, unique: true },
     password: { type: String, required: true },
 
-    role: { type: String, enum: ['user', 'admin'], default: 'user' },
+    role: { type: String, enum: ['user', 'admin', 'superAdmin'], default: 'user' },
 
     isActive: { type: Boolean, default: true },
     createdAt: { type: Date, default: Date.now},